Updated TestService with interface to expose only runTests method for executing testcases.
